Reorganize package contents
