+++ docs/source/notebooks/index.md
@@ -1,17 +1,11 @@
 # Examples
 
-:::{toctree}
-:maxdepth: 1
-
-continuous-contractual
-continuous-non-contractual
-mmm_example
-:::
 
 :::{toctree}
 :caption: CLVs
 :maxdepth: 1
 
-clv/gamma_gamma_dev
+mmm/mmm_example
+clv/bg-nbd
 clv/gamma_gamma
 :::
