+++ pymmmc/__init__.py
@@ -1,9 +1,3 @@
 __version__ = "0.0.1"
 
 from pymmmc import clv, mmm
-from pymmmc.distributions import ContContract, ContNonContract
-
-__all__ = [
-    "ContNonContract",
-    "ContContract",
-]
