+++ pymmmc/clv/__init__.py
@@ -1,11 +1,9 @@
-from pymmmc.clv import models
-from pymmmc.clv.models.basic import CLVModel
+from pymmmc.clv import distributions, models
 from pymmmc.clv.models.beta_geo import BetaGeoModel
 from pymmmc.clv.models.gamma_gamma import GammaGammaModel, GammaGammaModelIndividual
 
 __all__ = (
     "BetaGeoModel",
-    "CLVModel",
     "GammaGammaModel",
     "GammaGammaModelIndividual",
 )
