+++ pymmmc/mmm/utils.py
