--- pymmmc/tests/test_import.py
@@ -1,4 +1,0 @@
-def test_import_pymc():
-    import pymc as pm
-
-    assert pm.__version__ >= "4.0.0"
